Am adaugat Adresa, Comanda si ElementComanda

+++ DTOs/Produs/ProdusCreateDto.cs
@@ -8,7 +8,7 @@ namespace MyApi.DTOs.Produs
         [MaxLength(100)]
         public string Nume { get; set; } = string.Empty;
         [Range(0.01, 1000000)]
-        public double Pret {  get; set; }
+        public decimal Pret {  get; set; }
         [Range(1, int.MaxValue)]
         public int CategorieId { get; set; }
