+++ DTOs/Produs/ProdusDto.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public string? Nume { get; set; }
-        public double Pret { get; set; }
+        public decimal Pret { get; set; }
         public int CategorieId {  get; set; }
         public string? NumeCategorie { get; set; }
