+++ DTOs/Produs/ProdusFiltruDto.cs
@@ -3,8 +3,8 @@
     public class ProdusFiltruDto
     {
         public int? CategorieId { get; set; }
-        public double? PretMinim { get; set; }
-        public double? PretMaxim { get; set; }
+        public decimal? PretMinim { get; set; }
+        public decimal? PretMaxim { get; set; }
         public string? Nume { get; set; }
 
         public string? SortBy { get; set; }
